% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_bib.R
\name{clean_bib}
\alias{clean_bib}
\title{Clean BibTeX files}
\usage{
clean_bib(input_file, output_dir = NULL)
}
\arguments{
\item{input_file}{Character vector of path to .bib file}

\item{output_dir}{Directory to save cleaned files (default = same directory)}
}
\value{
Invisibly returns paths to cleaned files
}
\description{
Clean BibTeX files
}
\examples{
\dontrun{
clean_bib("myrefs.bib")
clean_bib(c("file1.bib", "file2.bib"))
}

}
